fix: animators may only be run on Looper threads
